Refactor `Corr.set_params`
All default values are now declared in the Corr.__init__ method. Each
subroutine modifies the defaults only of the attributes it needs, instead
of all available subroutines having to set the same 4 attributes.

Previously, the available subroutines would create an array  of the
possible combinations of the columns in the dataset (represented by
their index) using one of two distinct methods. Now, this operation has
been consolidated under the itertools.combinations() function by letting
each subroutine choose an appropiate value of the r parameter.

The original indexes array was created as a list, using a list
comprehension, and converted to a numpy.array. This step was replaced
with a faster, more efficient numpy.arange().

Other changes:

 - Rename both the method and some of its variables.
 - Ask the user to set the parameters related to their desired
   subroutine before asking them to set R-squared.