Create the title using f-strings
f-strings are more compact and readable than regular string
concatenations using the `+` operator.

Previously, the code would branch if it found a power n different than
zero. Since the only distinction between the two cases was the inclusion
of the power n, the conditional block was reduced to just one clause.