Replace structured numpy array with Pandas DataFrame
The use of a DataFrame allows for a more readable loop, and
a cleaner print statement.

As a side-benefit, pandas.DataFrame.sort_values has a built-in
ascending switch to reverse the order of the DataFrame. The easiness
with which one can reverse the order *will* enable the program to
present the results that are most important to the user first.